requirements.txt
python-telegram-bot
transformers
torch
datasets
accelerate
spacy
bitsandbytes
evaluate
sentencepiece
tokenizers
torchinfo
sacrebleu
rouge_score
peft
jiwer
wandb
python-dotenv
setuptools==59.5.0
safetensors
python-telegram-bot==13.15
transformers[sentencepiece]


